Create final draft
Hi! That's really great! Nice to see the logo come to life! 🖖

Also, please could you make sure that I get accredited for the logo (you basically re-assembled what I made) if I create the final version based on this input? 

I like it but it feels off because it was a bit rushed and off, and secondly I like to take full credit instead of doing 90% of the work only for someone else to remix it and run away with it 😉

Cheers!